zoom/requirements.txt
Django==1.8
Pillow==2.9.0
psycopg2==2.6.1
virtualenv==13.1.2
